docs: add module-level docs, CursorShape enum, Default impl, fix unwraps
pty.rs: add module-level doc and method-level doc comments
term.rs: define #[repr(u8)] CursorShape enum (Block/Underline/Beam)
         replacing raw u8 return; implement Default for CellInfo
cli/main.rs: add doc to build_concepts(); replace bare .unwrap()
         with .expect() for Ctrl+C handler and regex constructors
gdext/lib.rs: add doc comments to undocumented get_rows/get_cols

+++ crates/godopty-gdext/src/lib.rs
@@ -174,7 +174,7 @@ impl GodoptyTerminal {
     /// Cursor shape: 0 = Block, 1 = Underline, 2 = Beam.
     #[func]
     fn get_cursor_shape(&self) -> i64 {
-        self.with_grid(|g| g.cursor_shape() as i64, -1)
+        self.with_grid(|g| g.cursor_shape() as u8 as i64, -1)
     }
 
     /// Resize the terminal grid and PTY to `rows × cols`.
